add gates

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -108,6 +109,14 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+        $post = BlogPost::findOrFail($id);
+
+//        if(Gate::denies('update-post', $post)){
+//            abort(403);
+//        }
+
+        $this->authorize('update-post', $post);
+
         return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
     }
 
@@ -121,6 +130,13 @@ class PostController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+//        if(Gate::denies('update-post', $post)){
+//            abort(403);
+//        }
+
+        $this->authorize('update-post', $post);
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -139,6 +155,13 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+//        if(Gate::denies('delete-post', $post)){
+//            abort(403, "You can't delete this post!");
+//        }
+
+        $this->authorize('delete-post', $post);
+
         $post->delete();
 
         session()->flash('status', 'Post was deleted!');
